requirements.txt
torch~=2.0.1+cu117
torchvision~=0.15.2+cu117

uvicorn~=0.23.2
Pillow~=9.3.0
fastapi~=0.101.0
scikit-learn~=1.3.0